refactor(runner): decompose connection god-file into submodules
Re-applied at the crates/homeboy-core path after #8389 extracted the
homeboy-core crate.

connection.rs was ~3927 lines (god_file). Move the remote_daemon and
session_store inline mods into their own files, and split the oversized
test module into connection/tests/{recovery,session}.rs with shared test
helpers (direct_ssh_session, remote_daemon_status_for_test*) promoted to
connection/tests/mod.rs as pub(super).

connection.rs is now 1285 lines; every resulting file is under the
1500-line god_file threshold. cargo check -p homeboy-core --lib is clean.
No logic changed; public API preserved. Clears the god_file finding.
